Bug fix: RT#109159 Number formatting tests can fail
I believe the previous test was broken in that it was not anchored
nor was the use of . escaped, so it was foxed by some number
transformations.

This attempts - possibly futilely - to match the possible results out of
the number formatter for "reasonable" locales.   It may be necessary
to instead check the environment and skip if things look nasty!